[#166] Default 'httpMethod' to 'GET' to Ensure PSR-7 1.6.x Compatibility
PSR-7 requires a method to be set on Request objects at the time they
are created, but the Operation object previously defaulted the method to
an empty string. Now we default it to 'GET', which should be a safe
choice.

In addition, if a request is misconfigured to force the request method
to an empty string, there is now a validation for that.

Closes #166.