[#165] Add Support for Polymorphic Response Field Types
This allows fields within a response model to vary within a set of
allowed types instead of being required to be a specific JSON schema
value type.

This also corrects detection of null response values. It also fixes
detection of associative vs. indexed arrays so that the order in which
the type definitions appear in the service description doesn't affect
which type gets returned when a field can be either an object or an
array.

Closes #165.